+++ configure.py
@@ -630,7 +630,6 @@ scylla_tests = set([
     'test/unit/lsa_sync_eviction_test',
     'test/unit/row_cache_alloc_stress_test',
     'test/unit/row_cache_stress_test',
-    'test/unit/bptree_stress_test',
     'test/unit/btree_stress_test',
     'test/unit/bptree_compaction_test',
     'test/unit/btree_compaction_test',
@@ -1420,7 +1419,6 @@ tests_not_using_seastar_test_framework = set([
     'test/unit/lsa_async_eviction_test',
     'test/unit/lsa_sync_eviction_test',
     'test/unit/row_cache_alloc_stress_test',
-    'test/unit/bptree_stress_test',
     'test/unit/btree_stress_test',
     'test/unit/bptree_compaction_test',
     'test/unit/btree_compaction_test',
